feat: 골룸 단일 조회 타입 선언
